finish refactor
