+++ server/config/test.json
@@ -9,9 +9,6 @@
     "protocol": "http",
     "timeout": 3000
   },
-  "web3": {
-    "provider": "http://127.0.0.1:7545"
-  },
   "mongo": {
     "uri": "mongodb://localhost/CLN-community-app",
     "options": {}
