+++ server/src/events/tasks/bridge.js
@@ -5,7 +5,7 @@ const processPastEvents = require('./utils').processPastEvents
 const bridgeMapperABI = require('@constants/abi/BridgeMapper')
 
 const processPastBridgeMappingEvents = async () => {
-  const address = config.get('networks.home.addresses.BridgeMapper')
+  const address = config.get('network.home.addresses.BridgeMapper')
   const basicTokenContract = new web3.eth.Contract(bridgeMapperABI, address)
   const eventName = 'BridgeMappingUpdated'
   await processPastEvents(eventName, basicTokenContract)
