+++ server/src/utils/bridge.js
@@ -2,8 +2,8 @@ const config = require('config')
 const ForeignBridgeFactoryABI = require('@constants/abi/ForeignBridgeFactory.json')
 const HomeBridgeFactoryABI = require('@constants/abi/HomeBridgeFactory.json')
 const BridgeMapperABI = require('@constants/abi/BridgeMapper.json')
-const foreignAddressess = require('@utils/network').addresses
-const homeAddresses = config.get('networks.fuse.addresses')
+const foreignAddressess = config.get('network.foreign.addresses')
+const homeAddresses = config.get('network.home.addresses')
 const { fetchGasPrice, isZeroAddress } = require('@utils/network')
 const { handleReceipt } = require('@events/handlers')
 const home = require('@services/web3/home')
